Added Nav Struct
The Nav struct is a similar device used from my previous pedal designs.
It gives an easy method for setting up pseudo-multitasking with use of
switches and initialization confirmations.

For example, when starting the void loop() function, we can have some
code that only runs in the first loop:

  Nav n;
  void loop() {
    if ( n.not_init() ) {
      // code to set up something goes here.  only runs once.
      // once n.not_init() is run once, it sets the init flag.
    } else {
      // rest of code goes here
    }
  }

We can also reset the Nav struct to allow us to go back to the first if
statement when we want/need to.

The e() and jump_to() methods allows easy use of switches.  When calling
the jump_to() method, it resets not_init().  This allows setup for each
switch state:

  #define STATE_INIT 0
  #define STATE_INIT 0
  switch ( n.e() ) {
    case STATE_INIT:
      if ( n.not_init() ) {
        HW::board_led(ON);
      } else {
        if ( /* some button is hit */ ) n.jump_to(STATE_FIRST_STAGE);
      }
      break;

    case STATE_FIRST_STAGE:
      if ( n.not_init() ) {
        HW::board_led(OFF);
      } else {
        /* other code goes here */
      }
      break;
  };