Added custom ADC library
This library is set up as a single struct that we can use all of the
scattered ADC settings in one location.  Easily change settings via the
struct members.  Much easier to read variables and a lot of
documentation so we do not have to keep digging into the manual every 5
seconds.