Removed DMA
At this point of the process, if we end up using DMA, i think we will
end up using the DMAChannel.h file included with the teensy core
library.  This thing is a tad more complicated than I think we need for
this project.

Since this pedal will be operating on single samples at a time instead
of blocks of samples at a time, there really is no added benefit to
using DMA.