Successfully record audio to SD during audio transfer
So we are taking audio into an input buffer, then saving that buffer to
disk while the LED is on, then printing out the first buffer after the
LED turns off.  This is all written and read from the SD card in
real-time.

Still need to separate out the audio transfer from input-output vs SD
card writing.